irker-svnpoller: Allow limiting the amount of commit message lines sent to irker
This adds a MAX_COMMIT_LINES constant that controls the amount of lines
sent in the irker notification (default is 5). If there are more
non-empty lines in the commit message than the configured maximum, "(Log
message trimmed)" is sent at the end in a new line. As a side-effect, we
don't send empty lines anymore.